requirements.txt
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-crispy-forms==1.14.0
django-paypal==2.0
idna==3.3
Pillow==9.1.0
pytz==2018.7
requests==2.27.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9

